pyproject.toml
[project]
name = "latentlens"
version = "0.1.0"
description = "Streamlit-based visualization app."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "streamlit",
  "plotly",
  "umap-learn",
  "scikit-learn",
  "numpy",
  "pandas",
  "pinecone-client",
  "chromadb",
  "qdrant-client",
  "streamlit-plotly-events",
  "openai>=2.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Explicitly list packages because the project does not use a matching top-level
# package directory for auto-discovery.
packages = ["core", "utils"]
include = ["app.py"]
